requirements.txt
celery>=5.3.0
inflection==0.5.1
Jinja2==3.1.5
jmespath-community==1.1.2
jsonpointer==2.2
jsonschema==3.2.0
kombu>=5.3.0
openapi-schema-validator==0.1.5
prometheus-client==0.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.13
pyOpenSSL==20.0.1
python-dateutil>=2.8.2
python-string-utils==1.0.0
pytimeparse==1.1.8
PyYAML==6.0.1
redis>=4.5.0
requests==2.32.0
str2bool==1.1
StringGenerator==0.4.4
urllib3==1.26.19
